refactor: move the validate code to vcs package
