+++ plugin/db/driver_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -207,46 +206,5 @@ func TestParseMigrationInfo(t *testing.T) {
 			require.NoError(t, err)
 			require.Equal(t, tc.want, *mi)
 		})
-	}
-}
-
-func TestIsDoubleTimesAsteriskInTemplateValid(t *testing.T) {
-	tests := []struct {
-		template string
-		err      bool
-	}{
-		{
-			template: "**",
-			err:      true,
-		},
-		{
-			template: "bytebase/{{ENV_NAME}}/**",
-			err:      true,
-		},
-		{
-			template: "**/{{ENV_NAME}}/{{DB_NAME}}.sql",
-			err:      true,
-		},
-		{
-			template: "bytebase/**/{{ENV_NAME}}/**/{{DB_NAME}}__{{VERSION}}__{{TYPE}}__{{DESCRIPTION}}.sql",
-			err:      false,
-		},
-		{
-			template: "/**/{{ENV_NAME}}/**/{{DB_NAME}}__{{VERSION}}__{{TYPE}}__{{DESCRIPTION}}.sql",
-			err:      false,
-		},
-		// Credit to Linear Issue BYT-1267
-		{
-			template: "/configure/configure/{{ENV_NAME}}/**/**/{{DESCRIPTION}}.sql",
-			err:      false,
-		},
-	}
-	for _, test := range tests {
-		outputErr := isDoubleAsteriskInTemplateValid(test.template)
-		if test.err {
-			assert.Error(t, outputErr)
-		} else {
-			assert.NoError(t, outputErr)
-		}
 	}
 }
